Updated boolean example
The boolean query example used the query() method, which does not exist.  Instead, it should use the queryBuilder() method.  Additionally, if you pass a *single* parameter to must(), should(), or must_not().. then it will be iterated as though it is an array of class that implement \Sherlock\components\QueryInterface.

Therefore, you cannot pass a Term() or Match() class directly to the must(), chould(), or must_not() method as previously shown.